# speed up `hash(::Symbol)`