Since symbols are pre-hashed, add a fast path to return the value directly.

Before:
```
julia> @btime hash(:x)
  3.902 ns (0 allocations: 0 bytes)
julia> d = Dict(Symbol("x",i)=>i for i = 1:10000);
julia> @btime ($d)[:x6275]
  8.280 ns (0 allocations: 0 bytes)
```

After:
```
julia> @btime hash(:x)
  1.307 ns (0 allocations: 0 bytes)
julia> d = Dict(Symbol("x",i)=>i for i = 1:10000);
julia> @btime ($d)[:x6275]
  3.259 ns (0 allocations: 0 bytes)
```

Not sure why it speeds up dict lookup so much, but I'll take it.