@JeffBezanson 2021-06-01
I realized it's easy to keep the same `hash` values, which could be useful for tests etc. That keeps `Dict`s the same, at the expense of changing `IdDict`s. But, the nice part is that IdDicts with symbol keys will have the same order as Dicts.