@vtjnash 2021-06-03
julia-debug now fails when validating objectid
```
julia> code_llvm(objectid, (Symbol,), optimize=false, raw=true)
;  @ reflection.jl:302 within `objectid`
define i64 @julia_objectid_112({} addrspace(10)* nonnull %0) #0 !dbg !5 {
top:
  %x = alloca {} addrspace(10)*, align 8
  %1 = call {}*** @julia.get_pgcstack()
  store {} addrspace(10)* null, {} addrspace(10)** %x, align 8
  %2 = bitcast {}*** %1 to {}**
  %current_task = getelementptr inbounds {}*, {}** %2, i64 2305843009213693940
  %3 = bitcast {}** %current_task to i64*
  %world_age = getelementptr inbounds i64, i64* %3, i64 13
  store {} addrspace(10)* %0, {} addrspace(10)** %x, align 8
  %4 = load {} addrspace(10)*, {} addrspace(10)** %x, align 8, !dbg !7, !nonnull !4
  %5 = bitcast {} addrspace(10)* %4 to i64 addrspace(10)*, !dbg !7
  %6 = getelementptr inbounds i64, i64 addrspace(10)* %5, i64 2, !dbg !7
  %7 = load i64, i64 addrspace(10)* %6, align 8, !dbg !7, !tbaa !8, !invariant.load !4
  ret i64 %7, !dbg !7
}
GC tracked values may not appear in GEP expressions. You may have to decay the value first
  %6 = getelementptr inbounds i64, i64 addrspace(10)* %5, i64 2, !dbg !7

signal (6): Aborted
```